+++ line_profiler/line_profiler.py
@@ -5,7 +5,6 @@ inspect its output. This depends on the :py:mod:`line_profiler._line_profiler`
 Cython backend.
 """
 import pickle
-import functools
 import inspect
 import linecache
 import tempfile
@@ -20,6 +19,7 @@ except ImportError as ex:
         'The line_profiler._line_profiler c-extension is not importable. '
         f'Has it been compiled? Underlying error is ex={ex!r}'
     )
+from .profiler_mixin import ByCountProfilerMixin
 
 # NOTE: This needs to be in sync with ../kernprof.py and __init__.py
 __version__ = '4.3.0'
@@ -32,25 +32,7 @@ def load_ipython_extension(ip):
     ip.register_magics(LineProfilerMagics)
 
 
-def is_coroutine(f):
-    return inspect.iscoroutinefunction(f)
-
-
-CO_GENERATOR = inspect.CO_GENERATOR
-
-
-def is_generator(f):
-    """ Return True if a function is a generator.
-    """
-    isgen = (f.__code__.co_flags & CO_GENERATOR) != 0
-    return isgen
-
-
-def is_classmethod(f):
-    return isinstance(f, classmethod)
-
-
-class LineProfiler(CLineProfiler):
+class LineProfiler(CLineProfiler, ByCountProfilerMixin):
     """
     A profiler that records the execution times of individual lines.
 
@@ -73,85 +55,7 @@ class LineProfiler(CLineProfiler):
         it on function exit.
         """
         self.add_function(func)
-        if is_classmethod(func):
-            wrapper = self.wrap_classmethod(func)
-        elif is_coroutine(func):
-            wrapper = self.wrap_coroutine(func)
-        elif is_generator(func):
-            wrapper = self.wrap_generator(func)
-        else:
-            wrapper = self.wrap_function(func)
-        return wrapper
-
-    def wrap_classmethod(self, func):
-        """
-        Wrap a classmethod to profile it.
-        """
-        @functools.wraps(func)
-        def wrapper(*args, **kwds):
-            self.enable_by_count()
-            try:
-                result = func.__func__(func.__class__, *args, **kwds)
-            finally:
-                self.disable_by_count()
-            return result
-        return wrapper
-
-    def wrap_coroutine(self, func):
-        """
-        Wrap a Python 3.5 coroutine to profile it.
-        """
-
-        @functools.wraps(func)
-        async def wrapper(*args, **kwds):
-            self.enable_by_count()
-            try:
-                result = await func(*args, **kwds)
-            finally:
-                self.disable_by_count()
-            return result
-
-        return wrapper
-
-    def wrap_generator(self, func):
-        """ Wrap a generator to profile it.
-        """
-        @functools.wraps(func)
-        def wrapper(*args, **kwds):
-            g = func(*args, **kwds)
-            # The first iterate will not be a .send()
-            self.enable_by_count()
-            try:
-                item = next(g)
-            except StopIteration:
-                return
-            finally:
-                self.disable_by_count()
-            input_ = (yield item)
-            # But any following one might be.
-            while True:
-                self.enable_by_count()
-                try:
-                    item = g.send(input_)
-                except StopIteration:
-                    return
-                finally:
-                    self.disable_by_count()
-                input_ = (yield item)
-        return wrapper
-
-    def wrap_function(self, func):
-        """ Wrap a function to profile it.
-        """
-        @functools.wraps(func)
-        def wrapper(*args, **kwds):
-            self.enable_by_count()
-            try:
-                result = func(*args, **kwds)
-            finally:
-                self.disable_by_count()
-            return result
-        return wrapper
+        return self.wrap_callable(func)
 
     def dump_stats(self, filename):
         """ Dump a representation of the data to a file as a pickled LineStats
@@ -169,32 +73,6 @@ class LineProfiler(CLineProfiler):
         show_text(lstats.timings, lstats.unit, output_unit=output_unit,
                   stream=stream, stripzeros=stripzeros,
                   details=details, summarize=summarize, sort=sort, rich=rich)
-
-    def run(self, cmd):
-        """ Profile a single executable statment in the main namespace.
-        """
-        import __main__
-        main_dict = __main__.__dict__
-        return self.runctx(cmd, main_dict, main_dict)
-
-    def runctx(self, cmd, globals, locals):
-        """ Profile a single executable statement in the given namespaces.
-        """
-        self.enable_by_count()
-        try:
-            exec(cmd, globals, locals)
-        finally:
-            self.disable_by_count()
-        return self
-
-    def runcall(self, func, *args, **kw):
-        """ Profile a single function call.
-        """
-        self.enable_by_count()
-        try:
-            return func(*args, **kw)
-        finally:
-            self.disable_by_count()
 
     def add_module(self, mod):
         """ Add all the functions in a module and its classes.